+++ src/mame/tatsumi/tatsumi_m.cpp
@@ -18,111 +18,6 @@ void tatsumi_state::tatsumi_reset()
 
 /******************************************************************************/
 
-uint16_t roundup5_state::roundup_v30_z80_r(offs_t offset)
-{
-	address_space &targetspace = m_audiocpu->space(AS_PROGRAM);
-
-	/* Each Z80 byte maps to a V30 word */
-	if (m_control_word & 0x20)
-		offset += 0x8000; /* Upper half */
-
-	return 0xff00 | targetspace.read_byte(offset);
-}
-
-void roundup5_state::roundup_v30_z80_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	address_space &targetspace = m_audiocpu->space(AS_PROGRAM);
-
-	/* Only 8 bits of the V30 data bus are connected - ignore writes to the other half */
-	if (ACCESSING_BITS_0_7)
-	{
-		if (m_control_word & 0x20)
-			offset += 0x8000; /* Upper half of Z80 address space */
-
-		targetspace.write_byte(offset, data & 0xff);
-	}
-}
-
-
-void roundup5_state::roundup5_control_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	COMBINE_DATA(&m_control_word);
-
-	if (m_control_word & 0x10)
-		m_subcpu->set_input_line(INPUT_LINE_HALT, ASSERT_LINE);
-	else
-		m_subcpu->set_input_line(INPUT_LINE_HALT, CLEAR_LINE);
-
-	if (m_control_word & 0x4)
-		m_audiocpu->set_input_line(INPUT_LINE_HALT, ASSERT_LINE);
-	else
-		m_audiocpu->set_input_line(INPUT_LINE_HALT, CLEAR_LINE);
-
-//  if (offset == 1 && (tatsumi_control_w & 0xfeff) != (last_bank & 0xfeff))
-//      logerror("%s:  Changed bank to %04x (%d)\n", m_maincpu->pc(), tatsumi_control_w,offset);
-
-//todo - watchdog
-
-	/* Bank:
-
-	    0x0017  :   OBJ banks
-	    0x0018  :   68000 RAM       mask 0x0380 used to save bits when writing
-	    0x0c10  :   68000 ROM
-
-	    0x0040  :   Z80 rom (lower half) mapped to 0x10000
-	    0x0060  :   Z80 rom (upper half) mapped to 0x10000
-
-	    0x0080  :   enabled when showing map screen after a play
-	                (switches video priority between text layer and sprites)
-
-	    0x0100  :   watchdog.
-
-	    0x0c00  :   vram bank (bits 0x7000 also set when writing vram)
-
-	    0x8000  :   set whenever writing to palette ram?
-
-	    Changed bank to 0060 (0)
-	*/
-
-	if ((m_control_word & 0x8) == 0 && !(m_last_control & 0x8))
-		m_subcpu->set_input_line(INPUT_LINE_IRQ4, ASSERT_LINE);
-//  if (tatsumi_control_w&0x200)
-//      cpu_set_reset_line(1, CLEAR_LINE);
-//  else
-//      cpu_set_reset_line(1, ASSERT_LINE);
-
-//  if ((tatsumi_control_w&0x200) && (last_bank&0x200)==0)
-//      logerror("68k irq\n");
-//  if ((tatsumi_control_w&0x200)==0 && (last_bank&0x200)==0x200)
-//      logerror("68k reset\n");
-
-	if (m_control_word == 0x3a00)
-	{
-//      cpu_set_reset_line(1, CLEAR_LINE);
-//      logerror("68k on\n");
-	}
-
-	m_last_control = m_control_word;
-}
-
-void roundup5_state::road_vregs_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	/*
-	    ---- ---x ---- ---- enabled when there's a road slope of any kind, unknown purpose
-	    ---- ---- -xx- ---- enables alternatively in tunnels sometimes, color mods?
-	    ---- ---- ---x ---- road bank select
-	    ---- ---- ---- xxxx various values written during POST while accessing road pixel ram,
-	                        otherwise 0xb at the start of irq service
-	*/
-
-	COMBINE_DATA(&m_road_vregs[offset]);
-
-	m_subcpu->set_input_line(INPUT_LINE_IRQ4, CLEAR_LINE); // guess, probably wrong
-//  logerror("d_68k_e0000_w %s %04x\n", m_maincpu->pc(), data);
-}
-
-/******************************************************************************/
-
 void cyclwarr_state::cyclwarr_control_w(uint8_t data)
 {
 	m_control_word = data;
