+++ src/mame/tatsumi/tatsumi_v.cpp
@@ -96,347 +96,6 @@ TILE_GET_INFO_MEMBER(tatsumi_state::get_text_tile_info)
 
 /**********************************
  *
- * Round Up 5
- *
- *********************************/
-
-uint8_t roundup5_state::gfxdata_r(offs_t offset)
-{
-	if((m_control_word & 0x200) == 0x200)
-	{
-		offset += (m_control_word & 0x6000) << 2;
-
-		return m_bg_gfxram[offset];
-	}
-
-	offset+=((m_control_word&0x0c00)>>10) * 0x8000;
-	return m_tx_gfxram[offset];
-}
-
-void roundup5_state::gfxdata_w(offs_t offset, uint8_t data)
-{
-	if((m_control_word & 0x200) == 0x200)
-	{
-		offset += (m_control_word & 0x6000) << 2;
-		m_bg_gfxram[offset] = data;
-		return;
-	}
-
-	offset+=((m_control_word&0x0c00)>>10) * 0x8000;
-
-	if (offset>=0x18000 && data)
-		logerror("effective write to vram %06x %02x (control %04x)\n",offset,data,m_control_word);
-
-	m_tx_gfxram[offset] = data;
-
-	offset=offset%0x8000;
-
-	m_gfxdecode->gfx(0)->mark_dirty(offset/8);
-}
-
-VIDEO_START_MEMBER(roundup5_state,roundup5)
-{
-	m_tx_layer = &machine().tilemap().create(*m_gfxdecode, tilemap_get_info_delegate(*this, FUNC(tatsumi_state::get_text_tile_info)), TILEMAP_SCAN_ROWS, 8,8, 128,64);
-	m_tx_gfxram = std::make_unique<uint8_t[]>(0x20000);
-	m_bg_gfxram = std::make_unique<uint8_t[]>(0x20000);
-
-	m_tx_layer->set_transparent_pen(0);
-
-	m_gfxdecode->gfx(0)->set_source(m_tx_gfxram.get());
-
-	save_pointer(NAME(m_tx_gfxram), 0x20000);
-	save_pointer(NAME(m_bg_gfxram), 0x20000);
-}
-
-void roundup5_state::draw_road(bitmap_rgb32 &bitmap, const rectangle &cliprect)
-{
-/*
-0xf980 0x0008 0x8c80 0x4a00 - road right to below, width unknown (32 pixels guess)
-0xfa80 0x0008 0x8c80 0x4a00 - road right to below, width unknown (32 pixels guess)
-
-0xfb80 0x0008 0x8c80 0x4a00 - road in middle of screen, width unknown (32 pixels guess)
-
-0xfc80 0x0008 0x8c80 0x4a00 - road width above to left, same width as above (ie, xpos - 32)
-0xfd80 0x0008 0x8c80 0x4a00 - road width above to left, same width as above (ie, xpos - 32)
-0xfe80 0x0008 0x8c80 0x4a00 - road width above to left, same width as above (ie, xpos - 32)
-0xff80 0x0008 0x8c80 0x4a00 - road width above to left, same width as above (ie, xpos - 32)
-0x0001                      - road half/width above to left, (ie, xpos - 16)
-0x0081                      - road width to left as usual (xpos-16 from above, or 32 from above2)
-
-0xfb0b 0x210b 0x8cf5 0x0dea - blue & left & right, with  blue|----|----|----|----|blue
-in this mode changing columns 2 & 3 have no apparent effect
-0xfb0b 0x7b09 0x8cf5 0x0dea - as above, but scaled up - perhaps 18 pixels shifted (twice that overall size)
-0xfb0b 0x6c07 0x8cf5 0x0dea - as above, but scaled up - perhaps 40 pixels shifted from above
-0xfb0b 0xaa06 0x8cf5 0x0dea - as above, but scaled up - perhaps 16 pixels shifted from above
-0xfb0b 0xb005 0x8cf5 0x0dea - as above, but scaled up - perhaps 38 pixels shifted from above
-
-b21 diff is 1a6
-97b         20f
-76c         c2
-6aa         fa
-5b0
-
-
-0x0000 0x0008 0xxxxx 0xxxx - line starting at 0 for 128 pixels - 1 to 1 with road pixel data
-0xff00 0x0008 0xxxxx 0xxxx - line starting at 32 for 128 pixels - 1 to 1 with road pixel data
-0xfe00 0x0008 0xxxxx 0xxxx - line starting at 64 for 128 pixels - 1 to 1 with road pixel data
-
-
-
-at standard zoom (0x800)
-shift of 0x100 moves 32 pixels
-so shift of 8 is assumed to move 1 pixel
-
-at double zoom (0x1000)
-assume shift of 0x100 only moves 16 pixels
-so
-
-0x100 * 0x400 => 0x40
-0x100 * step 0x800 = must come out at 0x20
-0x100 * step 0x1000 = must come out at 0x10
-0x100 * step 0x2000 = 0x5
-
-pos is 11.5 fixed point
-
--0x580 = middle
--0x180
--0x080
-0
-0x80
-
-*/
-	const uint16_t *data = m_road_ctrl_ram;
-
-	// Road layer enable (?)
-	if ((m_vregs[0x1]&0x1)==0)
-		return;
-
-	// Road data bank select (double buffered)
-	if (m_road_vregs[0]&0x10)
-		data+=0x400;
-
-	// Apply clipping: global screen + local road y offsets
-	int y = 256 - ((m_vregs[0xa/2] >> 8) + m_road_yclip[0]);
-	data+=y*4;
-
-	int visible_line=0;
-
-	for ( ; y<cliprect.max_y+1; y++)
-	{
-		// TODO: tunnels road drawing has a different format?
-		// shift is always 0x88** while data[3] is a variable argument with bit 15 always on
-		int shift=data[0];
-		int shift2=data[2];
-		int pal = 4; //(data[3]>>8)&0xf;
-		int step = swapendian_int16(data[1]);
-		int samplePos=0;
-		uint16_t const *const linedata=m_road_pixel_ram;// + (0x100 * pal);
-		int startPos=0, endPos=0;
-
-		int palette_byte;//=m_road_color_ram[visible_line/8];
-
-		/*
-		    Each road line consists of up to two sets of 128 pixel data that can be positioned
-		    on the x-axis and stretched/compressed on the x-axis.  Any screen pixels to the left
-		    of the first set are drawn with pen 0 of the road pixel data.  Any screen pixels to the
-		    right of the second set line are drawn with pen 127 of the road pixel data.
-
-		    The road control data is laid out as follows (4 words per screen line, with 2 banks):
-
-		    Word 0: Line shift for 1st set - 13.3 signed fixed point value.
-		    Word 1: Line scale - 5.11 fixed point value.  So 0x800 is 1:1, 0x400 is 1:2, etc
-		    Word 2: Line shift for 2nd set - 13.3 signed fixed point value.
-		    Word 3: ?
-
-		    The scale is shared between both pixel sets.  The 2nd set is only used when the road
-		    forks into two between stages.  The 2nd line shift is an offset from the last pixel
-		    of the 1st set.  The 2nd line shift uses a different palette bank.
-
-2nd road uses upper palette - confirmed by water stage.
-offset is from last pixel of first road segment?
-//last pixel of first road is really colour from 2nd road line?
-
-		*/
-
-		palette_byte=m_road_color_ram[visible_line/8];
-		pal = 4 + ((palette_byte>>(visible_line%8))&1);
-
-		visible_line++;
-
-		if (shift&0x8000)
-			shift=-(0x10000 - shift);
-		shift=-shift;
-
-		if (step)
-			startPos=((shift<<8) + 0x80 )/ step;
-
-		int x;
-
-		/* Fill in left of road segment */
-		for (x=0; (x < startPos) && (x < cliprect.max_x+1); x++)
-		{
-			int col = linedata[0]&0xf;
-			bitmap.pix(y, x) = m_palette->pen(256 + pal*16 + col);
-		}
-
-		/* If startpos is negative, clip it and adjust the sampling position accordingly */
-		if (startPos<0)
-		{
-			samplePos=step*(0-startPos);
-			startPos=0;
-		}
-		else
-		{
-			samplePos=0;
-		}
-
-		/* Fill in main part of road, then right-hand side edge */
-		for (x=startPos; x < (cliprect.max_x + 1) && ((samplePos>>11)<0x80); x++)
-		{
-			// look up colour
-			int col = linedata[(samplePos>>11)&0x7f]&0xf;
-
-			/* Clamp if we have reached the end of the pixel data */
-			//if ((samplePos>>11) > 0x7f)
-			//  col=linedata[0x7f]&0xf;
-
-			bitmap.pix(y, x) = m_palette->pen(256 + pal*16 + col);
-
-			samplePos+=step;
-		}
-
-		/* Now work out how many pixels until start of 2nd segment */
-		startPos=x;
-
-		if (shift2&0x8000)
-			shift2=-(0x10000 - shift2);
-		shift2=-shift2;
-
-		if (step)
-			endPos=((shift2<<8) + 0x80) / step;
-		else
-			endPos=0;
-		endPos-=128;
-		endPos=startPos+endPos;
-
-		/* Fill pixels */
-		for (x=startPos; x < (cliprect.max_x+1) && (x < endPos); x++)
-		{
-			int col = linedata[0x80]&0xf;
-
-			/* Clamp if we have reached the end of the pixel data */
-			//if ((samplePos>>11) > 0x7f)
-			//  col=linedata[0x7f]&0xf;
-
-			bitmap.pix(y, x) = m_palette->pen(256 + pal*16 + col + 32);
-		}
-
-		if (endPos<0)
-		{
-			// end of left intersection (taking right turn)
-			samplePos=step*(0-startPos);
-		}
-		else if (endPos<x)
-		{
-			// start of right intersection
-			samplePos=step*(x-endPos);
-		}
-		else
-		{
-			// end of right intersection (taking right turn)
-			samplePos=0; // todo
-		}
-
-		for (/*x=endPos*/; x < cliprect.max_x+1; x++)
-		{
-			// look up colour
-			int col = linedata[((samplePos>>11)&0x7f) + 0x200]&0xf;
-
-			/* Clamp if we have reached the end of the pixel data */
-			if ((samplePos>>11) > 0x7f)
-				col=linedata[0x7f + 0x200]&0xf;
-
-			bitmap.pix(y, x) = m_palette->pen(256 + pal*16 + col + 32);
-
-			samplePos+=step;
-		}
-		data+=4;
-	}
-}
-
-// background layer landscape for Round Up 5
-// two bitmap layers, back layer is 512 x 128, the other one is 512 x 64
-// it's safe to assume that three monitor version will have a different arrangement here ...
-void roundup5_state::draw_landscape(bitmap_rgb32 &bitmap, const rectangle &cliprect, uint8_t type)
-{
-	// TODO: guess, assume back layer having less scroll increment than front for parallax scrolling.
-	// also notice that m_vregs[8/2] >> 8 is identical to [0x0c/2], always?
-	uint16_t x_base = type ? m_bg_scrollx[0] : m_vregs[0xc/2];
-	// TODO: maybe [0xa/2] applies here as well?
-	uint16_t y_base = m_bg_scrolly[0] & 0x1ff;
-	uint16_t y_scroll = 0x180 - y_base;
-	uint32_t base_offset;
-	uint16_t color_base = type ? 0x100 : 0x110;
-	int ysize = type ? 64 : 128;
-
-	base_offset = 0x10000 + type * 0x8000;
-	if(type)
-		y_scroll += 64;
-
-	//popmessage("%04x %04x %04x",m_vregs[8/2],m_vregs[0xc/2],m_bg_scrollx[0]);
-
-	for(int y = 0; y < ysize; y++)
-	{
-		for(int x = 0; x < 512; x++)
-		{
-			int res_x = (x_base + x) & 0x1ff;
-			uint32_t color = m_bg_gfxram[(res_x >> 1)+y*256+base_offset];
-
-			if(res_x & 1)
-				color >>= 4;
-
-			color &= 0xf;
-
-			if(cliprect.contains(x, y+y_scroll) && color)
-				bitmap.pix(y+y_scroll, x) = m_palette->pen(color+color_base);
-		}
-	}
-}
-
-uint32_t roundup5_state::screen_update_roundup5(screen_device &screen, bitmap_rgb32 &bitmap, const rectangle &cliprect)
-{
-	int tx_start_addr;
-
-	tx_start_addr = (m_hd6445_reg[0xc] << 8) | (m_hd6445_reg[0xd]);
-	tx_start_addr &= 0x3fff;
-
-	m_rotatingsprites->update_cluts();
-
-	m_tx_layer->set_scrollx(0,24);
-	m_tx_layer->set_scrolly(0,(tx_start_addr >> 4) | m_hd6445_reg[0x1d]);
-
-	bitmap.fill(m_palette->pen(384), cliprect); // todo
-	screen.priority().fill(0, cliprect);
-	m_rotatingsprites->draw_sprites(screen.priority(),cliprect,1,(m_sprite_control_ram[0xe0]&0x1000) ? 0x1000 : 0); // Alpha pass only
-	draw_landscape(bitmap,cliprect,0);
-	draw_landscape(bitmap,cliprect,1);
-	draw_road(bitmap,cliprect);
-	apply_shadow_bitmap(bitmap,cliprect,screen.priority(), 0);
-	if(m_control_word & 0x80) // enabled on map screen after a play
-	{
-		m_tx_layer->draw(screen, bitmap, cliprect, 0,0);
-		m_rotatingsprites->draw_sprites(bitmap,cliprect,0,(m_sprite_control_ram[0xe0]&0x1000) ? 0x1000 : 0); // Full pass
-	}
-	else
-	{
-		m_rotatingsprites->draw_sprites(bitmap,cliprect,0,(m_sprite_control_ram[0xe0]&0x1000) ? 0x1000 : 0); // Full pass
-		m_tx_layer->draw(screen, bitmap, cliprect, 0,0);
-	}
-	return 0;
-}
-
-/**********************************
- *
  * Cycle Warriors / Big Fight
  *
  *********************************/
